- Adding E2E test - Adding documentation - Renaming lighthouse container
